Big fix::
- Change the model to the normalized icosahedron
- Reset all the viewport to the origin
- Reset all the UI coordinate
- able to select mesh
- rotate works
- paint works

Problem::
- sculpting mode still have a problem due to the translation
- still have 8 color on the back buffer

+++ FingerTracking/picking.h
@@ -3,6 +3,6 @@
 
 //callback when picking occurs
 void picked(GLuint name, int sw);
-void processPick(int cursorX, int cursorY);
+void processPick(float cursorX, float cursorY);
 int getSelection();
 static int pickMe = 0;
